requirements.txt
streamlit
openai
faiss-cpu
pandas
numpy
openpyxl